--- BlogSiteMVC/Models/BaseEntity.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace BlogSiteMVC.Models
-{
-    public class BaseEntity
-    {
-        public int Id { get; set; }
-    }
-}
