--- BlogSiteMVC/Models/Topic.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace BlogSiteMVC.Models
-{
-    public class Topic : BaseEntity
-    {
-        public Topic()
-        {
-        }
-        public string Name { get; set; }
-        public int TotalReading { get; set; }
-    }
-}
