--- BlogSiteMVC/Repositories/Abstract/IBaseRepository.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace BlogSiteMVC.Repositories.Abstract
-{
-    interface IBaseRepository<T> 
-        where T : class
-    {
-        IEnumerable<T> GetAll();
-        T GetById(int id);
-        bool Add(T entity);
-        bool Update(T entity);
-        bool Delete(T entity);
-    }
-}
