--- BlogSiteMVC/Repositories/Abstract/IUserInformationRepository.cs
@@ -1,13 +1,0 @@
-﻿using BlogSiteMVC.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace BlogSiteMVC.Repositories.Abstract
-{
-    interface IUserInformationRepository : IBaseRepository<UserInformation>
-    {
-        IEnumerable<UserInformation> GetAllWithArticles();
-    }
-}
