+++ Business/Abstract/ITopicService.cs
@@ -1,0 +1,13 @@
+﻿using Core.Concrete;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Business.Abstract
+{
+    public interface ITopicService : IBaseService<Topic>
+    {
+    }
+}
