+++ Business/Abstract/IUserInformationService.cs
@@ -1,0 +1,14 @@
+﻿using Core.Concrete;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Business.Abstract
+{
+    public interface IUserInformationService : IBaseService<UserInformation>
+    {
+        IEnumerable<UserInformation> GetAllWithArticles();
+    }
+}
