+++ Core/Concrete/UserInformation.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace Core.Concrete
+{
+    public class UserInformation : BaseEntity
+    {
+        public UserInformation()
+        {
+            Articles = new HashSet<Article>();
+            Topics = new HashSet<Topic>();
+        }
+
+        public string FirstName { get; set; }
+        public string LastName { get; set; }
+        public byte[] Picture { get; set; }
+        public string Description { get; set; }
+        public int UserRegisterId { get; set; }
+        public UserRegister UserRegister { get; set; }
+
+        public ICollection<Article> Articles { get; set; }
+        public ICollection<Topic> Topics { get; set; }
+
+    }
+}
