+++ DataAccess/Abstract/IBaseRepository.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace DataAccess.Abstract
+{
+    interface IBaseRepository<T> 
+        where T : class
+    {
+        IEnumerable<T> GetAll();
+        T GetById(int id);
+        bool Add(T entity);
+        bool Update(T entity);
+        bool Delete(T entity);
+    }
+}
